requirements.txt
python-telegram-bot==22.3
deep-translator==1.11.4
pytz